% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_envirologger_data.R
\name{clean_envirologger_data}
\alias{clean_envirologger_data}
\title{Function to clean a data frame returned by \code{get_envirologger_data}.}
\usage{
clean_envirologger_data(df)
}
\arguments{
\item{df}{Data frame returned by \code{get_envirologger_data}.}
}
\description{
\code{clean_envirologger_data} creates a variable variable/column with help
of a look-up table and treats statuses as observations.
}
\seealso{
\code{\link{get_envirologger_data}}
}
\author{
Stuart K. Grange
}
