% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envirologger_sensors.R
\name{get_envirologger_sensors}
\alias{get_envirologger_sensors}
\title{Function to get available Envirologger API monitoring sensors.}
\usage{
get_envirologger_sensors(user, key)
}
\arguments{
\item{user}{An Envirologger API user-name.}

\item{key}{An Envirologger API key for \code{user}.}
}
\value{
Tibble.
}
\description{
Function to get available Envirologger API monitoring sensors.
}
\examples{
\dontrun{

# Get sensor information
get_envirologger_sensors(user, key)

}

}
\seealso{
\href{https://api.airmonitors.net/3.5/documentation}{API Documentation},
\code{\link{get_envirologger_data}}, \code{\link{get_envirologger_sensors}}
}
\author{
Stuart K. Grange
}
