% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_envirologger_xml.R
\name{read_envirologger_json}
\alias{read_envirologger_json}
\title{Function to read Envirologger XML documents which have been converted to JSON.}
\usage{
read_envirologger_json(file, tz = "UTC")
}
\arguments{
\item{file}{File name of Envirologger JSON file.}

\item{tz}{Time-zone of dates within \code{file}. Default is \code{"UTC"}.}
}
\description{
\code{read_envirologger_json} is faster than the XML version and is suitable
for much larger file sizes.
}
\examples{

\dontrun{

# Load a file
data_test <- read_envirologger_json("bt_tower_data.json")

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{read_envirologger_xml}}
}

