% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_envirologger_xml.R
\name{read_envirologger_xml}
\alias{read_envirologger_xml}
\title{Function to read Envirologger XML documents.}
\usage{
read_envirologger_xml(file, tz = "UTC")
}
\arguments{
\item{file}{File name of Envirologger XML file.}

\item{tz}{Time-zone of dates within \code{file}. Default is \code{"UTC"}.}
}
\description{
\code{read_envirologger_xml} is not fast and is not suitable for large XML 
documents.
}
\examples{

\dontrun{

# Load a file
data_test <- read_envirologger_xml("bt_tower_data.xml")

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{read_envirologger_json}}
}

