% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_label_look_up.R
\name{read_label_look_up}
\alias{read_label_look_up}
\title{Function to load look-up table for the Envirologger label and sensor names.}
\usage{
read_label_look_up()
}
\value{
Data frame with correct data types.
}
\description{
Function to load look-up table for the Envirologger label and sensor names.
}
\examples{
\dontrun{

# Load look-up table
data_labels <- read_label_look_up()

}

}
\author{
Stuart K. Grange
}
