% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envirologger_data.R
\name{get_envirologger_data}
\alias{get_envirologger_data}
\title{Function to get data from Envirologger API.}
\usage{
get_envirologger_data(user, key, station, server, start = NA, end = NA,
  tz = "UTC", extra = TRUE)
}
\arguments{
\item{user}{Envirologger API user-name.}

\item{key}{Envirologger API key for \code{user}.}

\item{station}{Station's data to download.}

\item{server}{API server to use for download.}

\item{start}{Start date of query.}

\item{end}{End date of query.}

\item{tz}{Time-zone for dates.}

\item{extra}{Should the returned data frame contain extra infomation? Default
is \code{TRUE} but setting to \code{FALSE} can be useful for packages such
as \strong{openair}.}
}
\description{
The Envirologger API is not fast.
}
\examples{
\dontrun{

# Get some data
data_air <- get_envirologger_data(user, key, 1000, 100, start = "2016-06-20", 
                                  end = "2016-06-21")

}

}
\author{
Stuart K. Grange
}
\seealso{
\href{http://api.envirologger.net/2.0/documentation}{Documentation},
\code{\link{get_envirologger_stations}}
}

