% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envirologger_data.R
\name{get_envirologger_data}
\alias{get_envirologger_data}
\title{Function to get observational data from Envirologger API.}
\usage{
get_envirologger_data(user, key, station, start = NA, end = NA,
  tz = "UTC", remove_duplicates = TRUE, interval = "24 hour",
  verbose = FALSE)
}
\arguments{
\item{user}{An Envirologger API user-name.}

\item{key}{An Envirologger API key for \code{user}.}

\item{station}{A vector of station codes to download. Station codes are 
integers and \code{\link{get_envirologger_stations}} can be used to find 
these codes.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{tz}{Time zone for the observations' dates. The default is \code{"UTC"}.}

\item{remove_duplicates}{Should "true" date-station-variable-value duplicates
be removed? Default is \code{TRUE} as this is common.}

\item{interval}{How much data should the function request from the API for 
each iteration?}

\item{verbose}{Should the functions give messages?}
}
\value{
Tibble.
}
\description{
Function to get observational data from Envirologger API.
}
\examples{

\dontrun{

# Get some data for a made up station
data_air <- get_envirologger_data(
  user = user, 
  key = key, 
  station = 1001, 
  start = "2016-06-20", 
  end = "2016-06-21"
)

}

}
\seealso{
\href{https://api.airmonitors.net/3.5/documentation}{API Documentation},
\code{\link{get_envirologger_stations}}
}
\author{
Stuart K. Grange
}
