% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envirologger_stations.R
\name{get_envirologger_stations}
\alias{get_envirologger_stations}
\title{Function to get avalaible Envirologger monitoring stations.}
\usage{
get_envirologger_stations(user, key)
}
\arguments{
\item{user}{Envirologger API user-name.}

\item{key}{Envirologger API key for \code{user}.}
}
\description{
Function to get avalaible Envirologger monitoring stations.
}
\examples{
\dontrun{

get_envirologger_stations()

}

}
\author{
Stuart K. Grange
}

