% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envirologger_stations.R
\name{get_envirologger_stations}
\alias{get_envirologger_stations}
\title{Function to get available Envirologger API monitoring stations for a user.}
\usage{
get_envirologger_stations(user, key)
}
\arguments{
\item{user}{An Envirologger API user-name.}

\item{key}{An Envirologger API key for \code{user}.}
}
\value{
Data frame with correct data types.
}
\description{
Function to get available Envirologger API monitoring stations for a user.
}
\examples{
\dontrun{

get_envirologger_stations(user, key)

}

}
\seealso{
\href{http://api.envirologger.net/2.0/documentation}{API Documentation},
\code{\link{get_envirologger_data}}
}
\author{
Stuart K. Grange
}
